MyJavaAppCourse25/src/test/java/ru/ermakov/MyJavaAppCourse25/service/AnnualBonusServiceImplTest.java
package ru.ermakov.MyJavaAppCourse25.service;

import org.junit.jupiter.api.Test;
import ru.ermakov.MyJavaAppCourse25.model.enums.Positions;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

class AnnualBonusServiceImplTest {

    @Test
    void calculate() {

        // given
        Positions positions = Positions.HR;
        double bonus = 2.0;
        int workDays = 243;
        double salary = 100000.00;

        //when
        double result = new AnnualBonusServiceImpl().calculate(positions, salary, bonus, workDays);

        //then
        double expected = 360493.8271604938;
        assertThat(result).isEqualTo(expected);
    }
}